+++ src/arch/util/machine-broadcast.c
@@ -4,9 +4,9 @@
 
 #include "spanningTree.h"
 
-CmiCommHandle CmiSendNetworkFunc(int destPE, int size, char *msg, int mode);
+CmiCommHandle CmiSendNetworkFunc(int destPE, size_t size, char *msg, int mode);
 
-static void handleOneBcastMsg(int size, char *msg) {
+static void handleOneBcastMsg(size_t size, char *msg) {
     CmiAssert(CMI_BROADCAST_ROOT(msg)!=0);
 #if CMK_OFFLOAD_BCAST_PROCESS
     if (CMI_BROADCAST_ROOT(msg)>0) {
@@ -49,7 +49,7 @@ static void processBcastQs(void) {
 #endif
 }
 
-static INLINE_KEYWORD void processProcBcastMsg(int size, char *msg) {
+static INLINE_KEYWORD void processProcBcastMsg(size_t size, char *msg) {
     /* Since this function is only called on intermediate nodes,
      * the rank of this msg should be 0.
      */
@@ -76,7 +76,7 @@ static INLINE_KEYWORD void processProcBcastMsg(int size, char *msg) {
 
 
 #if CMK_NODE_QUEUE_AVAILABLE
-static INLINE_KEYWORD void processNodeBcastMsg(int size, char *msg) {
+static INLINE_KEYWORD void processNodeBcastMsg(size_t size, char *msg) {
 #if CMK_BROADCAST_SPANNING_TREE
     SendSpanningChildrenNode(size, msg);
 #elif CMK_BROADCAST_HYPERCUBE
@@ -94,7 +94,7 @@ static INLINE_KEYWORD void processNodeBcastMsg(int size, char *msg) {
 }
 #endif
 
-static void SendSpanningChildren(int size, char *msg, int rankToAssign, int startNode) {
+static void SendSpanningChildren(size_t size, char *msg, int rankToAssign, int startNode) {
 #if CMK_BROADCAST_SPANNING_TREE
     int i, oldRank;
     char *newmsg;
@@ -145,7 +145,7 @@ static void SendSpanningChildren(int size, char *msg, int rankToAssign, int star
 #endif
 }
 
-static void SendHyperCube(int size,  char *msg, int rankToAssign, int startNode) {
+static void SendHyperCube(size_t size,  char *msg, int rankToAssign, int startNode) {
 #if CMK_BROADCAST_HYPERCUBE
     int i, cnt, tmp, relDist, oldRank;
     const int dims=CmiNodesDim;
@@ -194,7 +194,7 @@ static void SendHyperCube(int size,  char *msg, int rankToAssign, int startNode)
 #endif
 }
 
-static void SendSpanningChildrenProc(int size, char *msg) {
+static void SendSpanningChildrenProc(size_t size, char *msg) {
     int startnode = CMI_BROADCAST_ROOT(msg)-1;
     SendSpanningChildren(size, msg, 0, startnode);
 #if CMK_SMP
@@ -204,7 +204,7 @@ static void SendSpanningChildrenProc(int size, char *msg) {
 }
 
 /* send msg along the hypercube in broadcast. (Sameer) */
-static void SendHyperCubeProc(int size, char *msg) {
+static void SendHyperCubeProc(size_t size, char *msg) {
     int startpe = CMI_BROADCAST_ROOT(msg)-1;
     int startnode = CmiNodeOf(startpe);
 #if CMK_SMP
@@ -218,11 +218,11 @@ static void SendHyperCubeProc(int size, char *msg) {
 }
 
 #if CMK_NODE_QUEUE_AVAILABLE
-static void SendSpanningChildrenNode(int size, char *msg) {
+static void SendSpanningChildrenNode(size_t size, char *msg) {
     int startnode = -CMI_BROADCAST_ROOT(msg)-1;
     SendSpanningChildren(size, msg, DGRAM_NODEMESSAGE, startnode);
 }
-static void SendHyperCubeNode(int size, char *msg) {
+static void SendHyperCubeNode(size_t size, char *msg) {
     int startnode = -CMI_BROADCAST_ROOT(msg)-1;
     SendHyperCube(size, msg, DGRAM_NODEMESSAGE, startnode);
 }
@@ -230,7 +230,7 @@ static void SendHyperCubeNode(int size, char *msg) {
 
 #if USE_COMMON_SYNC_BCAST
 /* Functions regarding broadcat op that sends to every one else except me */
-void CmiSyncBroadcastFn1(int size, char *msg) {
+void CmiSyncBroadcastFn1(size_t size, char *msg) {
     int i, mype;
 
     CQdCreate(CpvAccess(cQdState), CmiNumPes()-1);
@@ -264,7 +264,7 @@ void CmiSyncBroadcastFn1(int size, char *msg) {
     /*CmiPrintf("In  SyncBroadcast broadcast\n");*/
 }
 
-void CmiSyncBroadcastFn(int size, char *msg) {
+void CmiSyncBroadcastFn(size_t size, char *msg) {
     char *newmsg = msg;
 #if CMK_BROADCAST_SPANNING_TREE && CMK_BROADCAST_USE_CMIREFERENCE
       /* need to copy the msg in case the msg is on the stack */
@@ -277,7 +277,7 @@ void CmiSyncBroadcastFn(int size, char *msg) {
 #endif
 }
 
-void CmiFreeBroadcastFn(int size, char *msg) {
+void CmiFreeBroadcastFn(size_t size, char *msg) {
     CmiSyncBroadcastFn1(size,msg);
     CmiFree(msg);
 }
@@ -287,7 +287,7 @@ void CmiFreeBroadcastFn(int size, char *msg) {
 
 #if USE_COMMON_ASYNC_BCAST
 /* FIXME: should use spanning or hypercube, but luckily async is never used */
-CmiCommHandle CmiAsyncBroadcastFn(int size, char *msg) {
+CmiCommHandle CmiAsyncBroadcastFn(size_t size, char *msg) {
     /*CmiPrintf("In  AsyncBroadcast broadcast\n");*/
     CmiAbort("CmiAsyncBroadcastFn should never be called");
     return 0;
@@ -295,7 +295,7 @@ CmiCommHandle CmiAsyncBroadcastFn(int size, char *msg) {
 #endif
 
 /* Functions regarding broadcat op that sends to every one */
-void CmiSyncBroadcastAllFn(int size, char *msg) {
+void CmiSyncBroadcastAllFn(size_t size, char *msg) {
     char *newmsg = msg;
 #if CMK_BROADCAST_SPANNING_TREE && CMK_BROADCAST_USE_CMIREFERENCE
       /* need to copy the msg in case the msg is on the stack */
@@ -309,7 +309,7 @@ void CmiSyncBroadcastAllFn(int size, char *msg) {
 #endif
 }
 
-void CmiFreeBroadcastAllFn(int size, char *msg) {
+void CmiFreeBroadcastAllFn(size_t size, char *msg) {
     CmiSyncBroadcastFn1(size, msg);
 #if CMK_BROADCAST_SPANNING_TREE && CMK_BROADCAST_USE_CMIREFERENCE
       /* need to copy the msg in case the msg is on the stack */
@@ -323,14 +323,14 @@ void CmiFreeBroadcastAllFn(int size, char *msg) {
     CmiSendSelf(msg);
 }
 
-CmiCommHandle CmiAsyncBroadcastAllFn(int size, char *msg) {
+CmiCommHandle CmiAsyncBroadcastAllFn(size_t size, char *msg) {
     CmiSendSelf(CopyMsg(msg, size));
     return CmiAsyncBroadcastFn(size, msg);
 }
 
 #if CMK_NODE_QUEUE_AVAILABLE
 #if USE_COMMON_SYNC_BCAST
-void CmiSyncNodeBroadcastFn(int size, char *msg) {
+void CmiSyncNodeBroadcastFn(size_t size, char *msg) {
     int mynode = CmiMyNode();
     int i;
     CQdCreate(CpvAccess(cQdState), CmiNumNodes()-1);
@@ -348,30 +348,30 @@ void CmiSyncNodeBroadcastFn(int size, char *msg) {
 #endif
 }
 
-void CmiFreeNodeBroadcastFn(int size, char *msg) {
+void CmiFreeNodeBroadcastFn(size_t size, char *msg) {
     CmiSyncNodeBroadcastFn(size, msg);
     CmiFree(msg);
 }
 #endif
 
 #if USE_COMMON_ASYNC_BCAST
-CmiCommHandle CmiAsyncNodeBroadcastFn(int size, char *msg) {
+CmiCommHandle CmiAsyncNodeBroadcastFn(size_t size, char *msg) {
     CmiSyncNodeBroadcastFn(size, msg);
     return 0;
 }
 #endif
 
-void CmiSyncNodeBroadcastAllFn(int size, char *msg) {
+void CmiSyncNodeBroadcastAllFn(size_t size, char *msg) {
     CmiSyncNodeSendFn(CmiMyNode(), size, msg);
     CmiSyncNodeBroadcastFn(size, msg);
 }
 
-CmiCommHandle CmiAsyncNodeBroadcastAllFn(int size, char *msg) {
+CmiCommHandle CmiAsyncNodeBroadcastAllFn(size_t size, char *msg) {
     CmiSendNodeSelf(CopyMsg(msg, size));
     return CmiAsyncNodeBroadcastFn(size, msg);
 }
 
-void CmiFreeNodeBroadcastAllFn(int size, char *msg) {
+void CmiFreeNodeBroadcastAllFn(size_t size, char *msg) {
     CmiSyncNodeBroadcastFn(size, msg);
     /* Since it's a node-level msg, the msg could be executed on any other
      * procs on the same node. This means, the push of this msg to the
@@ -387,17 +387,17 @@ void CmiFreeNodeBroadcastAllFn(int size, char *msg) {
 
 #if ! CMK_MULTICAST_LIST_USE_COMMON_CODE
 
-void CmiSyncListSendFn(int npes, int *pes, int len, char *msg)
+void CmiSyncListSendFn(int npes, int *pes, size_t len, char *msg)
 {
     LrtsSyncListSendFn(npes, pes, len, msg);
 }
 
-CmiCommHandle CmiAsyncListSendFn(int npes, int *pes, int len, char *msg)
+CmiCommHandle CmiAsyncListSendFn(int npes, int *pes, size_t len, char *msg)
 {
     return LrtsAsyncListSendFn(npes, pes, len, msg);
 }
 
-void CmiFreeListSendFn(int npes, int *pes, int len, char *msg)
+void CmiFreeListSendFn(int npes, int *pes, size_t len, char *msg)
 {
     LrtsFreeListSendFn(npes, pes, len, msg);
 }
